SBMetadataHelper: hack path to Perl parser
Method parseFilename: relies on an external parser written in Perl to
make sense of movie and TV show file names. In Subler, this parser is
installed as part of the application bundle. In tagler, we don't have a
bundle, hence we need to divise another way of referencing it.

We try $HOME/Library/Application Support/tagler/ first, but if it
doesn't exist, we read the "SBMetadataPreference|PerlPath" property.

From the command line, set the property using
  defaults write tagler "SBMetadataPreference|PerlPath" path/to/scripts
To read back the defaults, use
  defaults read tagler

This should work well enough for now.