main.m: move file name parsing code to make it conditional
We now skip parsing the file name to derive title, season and episode
if all three are passed in from the command line. We continue to parse
the file name if only one or two of these are passed in. That way, it
is possible to override some properties while retrieving the others
from the file name.

This also means that we always have to parse movie file names, since we
don't know at that point whether they are, in fact, movies, or whether
the user just meant to override the TV show name, but derive season and
episode from the file name. We only know after parsing the name which
case it was.