main.m: added option to preserve timestamps (-P)
Using the command line option -P, it is possible to preserve the
timestamps of the video files being processed. This may be helpful for
bulk updates (such as changing genre, title art or total number of
tracks in a season) without having all modified videos show up as new
files.

Beware, however, that backup software relying on timestamps may not
recognize the files as having been modified and may thus not back them
up.