MetadataImporters: implement respective other search functions
We implement searchTVSeries: in SBTheMovieDB3 and searchMovie: in
SBTheTVDB and SBTheTVDB2. These methods all return nil.

The point of implementing them is to prevent a run time crash if the
wrong search method is sent to an object (i.e. requesting a TV series
search from a movie object), which can happen if the "-p" command line
option is being used.

Returning nil (i.e. no search result) is the proper solution (asking a
movie object to return a TV series will obviously result in no hits).