SBTheTVDB2.m: handle multiple search results
Searching for a TV series may yield many search results, not just the
show we were trying to find. Also, the results may be in arbitrary
order. There is no guarantee that the best match to the search term is
the first result.

Consequently, we have to iterate over all search results to find the
one we want.